SKCC: 'de-emphasise' the SKCC freqs outside the RSGB bandplans
Some of the SKCC frequencies are not very suitable for RSGB/UK or
Region1 use - they were originally defined for the USA, and don't
align well with the useable regions. There are alternaive SKCC
frequencies that do align, so we'll emphasise the good ones by
leaving them in CAPITALS and de-emphasise the less suitable ones
by moving them to 'lower case'. Frequencies in question are:

 7055
 7120
 14114
 21114

We leave 28114 alone, as despite 'feeling' like it might be errant,
it does fall into a reasonable narrow band.

Bump version to v1.3 as well.

Signed-off-by: Graham Whaley <graham.whaley@gmail.com>